+++ src/senaite/storage/browser/container/samples.py
@@ -22,10 +22,12 @@ import collections
 
 from bika.lims import api
 from bika.lims import senaiteMessageFactory as _s
+from plone.memoize.view import memoize
 from senaite.app.listing.view import ListingView
 from senaite.core.api import dtime
 from senaite.core.catalog import SAMPLE_CATALOG
 from senaite.storage import senaiteMessageFactory as _
+from senaite.storage import api as sapi
 from senaite.storage.permissions import TransitionAddSamples
 
 
@@ -76,8 +78,19 @@ class SampleListingView(ListingView):
             ("getDateReceived", {
                 "title": _s("Date Received"),
                 "toggle": True}),
+            ("getDateSored", {
+                "title": _(
+                    u"listing_samples_column_date_stored",
+                    default=u"Date stored"
+                ),
+                "index": "getDateStored",
+                "toggle": True}),
             ("getStorageExpiryDate", {
-                "title": _("Storage Expiry Date"),
+                "title": _(
+                    u"listing_samples_column_storage_expiry_date",
+                    default=u"Retain Until"
+                ),
+                "index": "getStorageExpiryDate",
                 "toggle": True}),
             ("Client", {
                 "title": _s("Client"),
@@ -133,6 +146,13 @@ class SampleListingView(ListingView):
         """
         return self.context.plone_utils.addPortalMessage(message, level)
 
+    @memoize
+    def get_warning_days_before_expiration(self):
+        """Returns the number of days before a sample's retention period ends
+        when it should be marked as approaching expiration
+        """
+        return sapi.get_warning_days_before_expiration()
+
     def folderitems(self):
         """We add this function to tell baselisting to use brains instead of
         full objects"""
@@ -156,15 +176,18 @@ class SampleListingView(ListingView):
             item["PreviousState"] = self.translate_review_state(
                 prev_state, api.get_portal_type(obj))
 
-        # storage expiry date
+        # date when the retention period expires
         column = "getStorageExpiryDate"
         expiry_date = obj.getStorageExpiryDate()
-        item[column] = self.ulocalized_time(expiry_date)
+        expiry_str = dtime.to_localized_time(expiry_date)
+        item[column] = expiry_str
 
-        # display in red if retention expired
-        if expiry_date <= dtime.DateTime():
-            expiry = self.ulocalized_time(expiry_date)
-            span = "<span class='text-danger'>%s</span>" % expiry
-            item["replace"][column] = span
+        # visual indicators if retention expired or approaching expiration
+        warn_days = self.get_warning_days_before_expiration()
+        span = "<span class='font-weight-bold %s'>%s</span>"
+        if sapi.is_retention_expired(obj):
+            item["replace"][column] = span % ("text-danger", expiry_str)
+        elif sapi.is_retention_approaching_expiration(obj, warn_days):
+            item["replace"][column] = span % ("text-warning", expiry_str)
 
         return item